Pretty sure 3-1 is alr correct

+++ function-3-1.cpp
@@ -11,7 +11,7 @@ bool is_fanarray(int array[], int n){
             }
         }
         else{
-            int middle=(int) ((n/2)-0.5);
+            int middle=(int) (((double)n)/2-0.5); // changed
             for (int i=0; i< middle ; i++){
                 if((array[i]!=array[n-1-i])||(array[i]>array[i+1])){
                     ans=false;
